sqlparser: generate faster AST formatting code
